Libre: load code-intel schema
Although there are no actual resolvers for these schemas available, since they
are enterprise code we don't use/have, the availability of the schema ensures
the actual code-search is actually usable.

The real fix, as the FIXME indicates, is to fix the call-sites to not require
the otherwise missing schemas.